[FLINK-21485][sql-client] Simplify the ExecutionContext
