+++ flink-table/flink-sql-client/src/test/resources/META-INF/services/org.apache.flink.table.factories.TableFactory
@@ -19,4 +19,4 @@ org.apache.flink.table.client.gateway.utils.SimpleCatalogFactory
 org.apache.flink.table.client.gateway.local.DependencyTest$TestCatalogFactory
 org.apache.flink.table.client.gateway.local.DependencyTest$TestHiveCatalogFactory
 org.apache.flink.table.client.gateway.local.DependencyTest$TestModuleFactory
-org.apache.flink.table.client.gateway.local.ExecutionContextTest$TestClassLoaderCatalogFactory
+org.apache.flink.table.client.gateway.context.ExecutionContextTest$TestClassLoaderCatalogFactory
